requirements.txt
Flask==1.1.2
Flask-JWT-Extended==3.25.0
Flask-RESTful==0.3.8
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2

flask-marshmallow==0.14.0
#flask-restx==0.2.0
python-dotenv==0.15.0
PyMySQL==0.10.1


